Move VaultMoc to make_mocs.py
